renamed modal folder to alert
